[CHEF-3552] Enable whyrun support for remote_directory provider.
The remote_directory provider calls functions from the directory
provider which is whyrun enabled.  This results in the state-changing
actions(namely mkdir) in Chef::Provider::Directory#action_create to
actually be called *after* the actions in the
Chef::Provider::RemoteDirecotry#action_create function, even if super
is called at the beginning of the latter function.